Cuisine page works with restrictions for moving to the next page, still needs to be refactored so vertical carousal is reusabe
