+++ components/QuizPages/VerticalCarousel.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import {
   Carousel,
   CarouselContent,
@@ -20,30 +20,23 @@ const cuisineOptions: CuisineOption[] = [
   { id: 6, name: "Pasta", image: "/pasta.jpg" },
 ];
 
-const VerticalCarousel = () => {
-  const [selectedCuisines, setSelectedCuisines] = useState<number[]>([]);
+interface VerticalCarouselProps {
+  selectedCuisines: number[];
+  onSelect: (cuisineId: number) => void;
+}
 
-  const handleSelection = (cuisineId: number) => {
-    setSelectedCuisines((prev) => {
-      //if already selected, it gets deselected
-      if (prev.includes(cuisineId)) {
-        return prev.filter((id) => id !== cuisineId);
-      }
-      // cannot be more than three options selected
-      if (prev.length < 3) {
-        return [...prev, cuisineId];
-      }
-      return prev;
-    });
-  };
-
+const VerticalCarousel = ({
+  selectedCuisines,
+  //setSelectedCuisines,
+  onSelect,
+}: VerticalCarouselProps) => {
   return (
     <Carousel
       orientation="vertical"
       opts={{
         axis: "y",
         dragFree: true,
-        containScroll: false, //allows free dragging without snapping back
+        containScroll: false,
         align: "start",
       }}
       className="w-full relative h-full overflow-hidden px-6"
@@ -53,12 +46,12 @@ const VerticalCarousel = () => {
           <CarouselItem key={cuisine.id} className="basis-auto">
             <button
               className={`w-full overflow-hidden relative rounded-2xl transition-all duration-200 ease-in-out
-    ${
-      selectedCuisines.includes(cuisine.id)
-        ? "ring-2 ring-red-500 scale-[0.85] hover:scale-[0.90]" // Scale down when selected, slightly bigger on hover
-        : "scale-100 hover:scale-105"
-    }`} // Normal state with hover effect
-              onClick={() => handleSelection(cuisine.id)}
+                ${
+                  selectedCuisines.includes(cuisine.id)
+                    ? "ring-2 ring-red-500 scale-[0.85] hover:scale-[0.90]"
+                    : "scale-100 hover:scale-105"
+                }`}
+              onClick={() => onSelect(cuisine.id)}
               aria-pressed={selectedCuisines.includes(cuisine.id)}
             >
               <div className="aspect-square relative">
@@ -67,11 +60,7 @@ const VerticalCarousel = () => {
                   alt={cuisine.name}
                   className="object-cover w-full h-full"
                 />
-                {/* Base overlay - always present */}
-                <div
-                  className={`absolute inset-0 bg-gradient-to-t from-black/60 to-transparent`}
-                />
-                {/* Red selection overlay - only shows when selected */}
+                <div className="absolute inset-0 bg-gradient-to-t from-black/60 to-transparent" />
                 <div
                   className={`absolute inset-0 bg-gradient-to-t from-red-500/80 to-transparent
                     transition-opacity duration-200 ease-in-out
